Implement Practice01
In watching the lecture video, practice01 is done.

The 'grid-template-columns' means how to define the column
in grid. This value can be 'repeat()' and % value, etc.

The 'grid-row' and 'grid-column' can be used for position and
size. The 'span' keyword means size.
'1 / 3' means that from 1 to 2 (<= <).
'1 / span 2' means that from 1, size is 2.
'span 2 / 3' means that size is 2, to 3.

See also: #CSS-1